add solution of rnnlm

+++ assignment2/data_utils/ner.py
@@ -5,9 +5,10 @@
 
 from utils import invert_dict
 from numpy import *
+import numpy as np
 
 def load_wv(vocabfile, wvfile):
-    wv = loadtxt(wvfile, dtype=float)
+    wv = loadtxt(wvfile, dtype=np.float32)
     with open(vocabfile) as fd:
         words = [line.strip() for line in fd]
     num_to_word = dict(enumerate(words))
